move user service to application
